Record user input actions automatically
